manda-processing/pyproject.toml
[project]
name = "manda-processing"
version = "0.1.0"
description = "Document processing service for Manda platform"
requires-python = ">=3.12"
readme = "README.md"
license = { text = "Proprietary" }
authors = [
    { name = "Manda Team" }
]

dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",

    # Database
    "asyncpg>=0.30.0",
    "supabase>=2.10.0",
    "neo4j>=5.15.0",  # Neo4j graph database driver (E4.15)

    # Graphiti - Temporal Knowledge Graph (E10.1, E10.2)
    "graphiti-core[google-genai,voyageai]>=0.3.0",  # Google Gemini for entity extraction + VoyageAI for embeddings

    # Document parsing (E3.2: Docling Integration)
    "docling>=2.15.0",
    "openpyxl>=3.1.5",
    "python-docx>=1.1.2",
    "tiktoken>=0.9.0",  # Token counting for chunking

    # AI/ML - Embeddings (E3.4) and LLM Analysis (E3.5)
    "openai>=1.82.0",
    "langchain>=0.3.0",
    "langchain-google-genai>=2.1.0",
    "langchain-core>=0.3.0",

    # Google Cloud
    "google-cloud-storage>=2.19.0",

    # Utilities
    "httpx>=0.28.0",
    "structlog>=24.4.0",
    "tenacity>=9.0.0",
    "pyyaml>=6.0.0",  # E11.6: Model config YAML loading

    # Pydantic AI (E11.5: Type-safe agent tools)
    "pydantic-ai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "logfire>=3.0.0",  # E11.5: Optional observability for Pydantic AI
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --tb=short"
markers = [
    "integration: marks tests as integration tests (may require external services)",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
